html/semantics/forms/the-input-element/radio.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>input type radio</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/#radio-button-state-(type=radio)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input type=radio name=group1 id=radio1>
<input type=radio name=group1 id=radio2>

<input type=radio name=groüp2 id=radio3>
<input type=radio name=groüp2 id=radio4>

<input type=radio id=radio5>
<input type=radio id=radio6 disabled>

<input type=radio name="group5" id=radio71 checked>
<input type=radio name="group5" id=radio72>

<input type=radio name=group3 id=radio8 checked>
<input type=radio name=group3 id=radio9>
<input type=radio name=group4 id=radio10>
<input type=radio name=group4 id=radio11 checked>

<form id="testform"></form>
<input type=radio form=testform name=group6 id=radio12 checked>
<input type=radio form=testform name=group6 id=radio13>
<input type=radio form=testform name=group6 id=radio14>

<form>
  <input type=radio name=group7 id=radio151 checked>
  <input type=radio name=group7 id=radio152>
</form>
<form>
  <input type=radio name=group7 id=radio153 checked>
  <input type=radio name=group7 id=radio154>
</form>
<input type=radio name=group7 id=radio155 checked>
<input type=radio name=group7 id=radio156>

<input type=radio name=group8 id=radio161 checked>
<input type=checkbox name=group8 id=radio162 checked>
<input type=radio name=group8 id=radio163>

<input type=radio id=radio171 checked>
<input type=radio id=radio172>

<div id='required'>
  <input type=radio name=group9 id=radio181 required>
  <input type=radio name=group9 id=radio182>
</div>

<input type=radio name=group10 id=radio191 checked>
<input type=checkbox name=group10 id=radio192 checked>
<input type=radio name=group10 id=radio193>

<form id="testform201"></form>
<form id="testform202"></form>
<input type=radio name=group11 form=testform201 id=radio201 checked>
<input type=radio name=group11 form=testform202 id=radio202 checked>
<input type=radio name=group11 form=testform202 id=radio203>

<script>
  var radio1 = document.getElementById('radio1'),
      radio2 = document.getElementById('radio2'),
      radio3 = document.getElementById('radio3'),
      radio4 = document.getElementById('radio4'),
      radio5 = document.getElementById('radio5'),
      radio6 = document.getElementById('radio6'),
      radio71 = document.getElementById('radio71'),
      radio72 = document.getElementById('radio72'),
      radio8 = document.getElementById('radio8'),
      radio9 = document.getElementById('radio9'),
      radio10 = document.getElementById('radio10'),
      radio11 = document.getElementById('radio11'),
      radio12 = document.getElementById('radio12'),
      radio13 = document.getElementById('radio13'),
      radio14 = document.getElementById('radio14'),
      radio151 = document.getElementById('radio151'),
      radio152 = document.getElementById('radio152'),
      radio153 = document.getElementById('radio153'),
      radio154 = document.getElementById('radio154'),
      radio155 = document.getElementById('radio155'),
      radio156 = document.getElementById('radio156'),
      radio161 = document.getElementById('radio161'),
      radio162 = document.getElementById('radio162'),
      radio163 = document.getElementById('radio163'),
      radio171 = document.getElementById('radio171'),
      radio172 = document.getElementById('radio172'),
      radio181 = document.getElementById('radio181'),
      radio182 = document.getElementById('radio182'),
      radio191 = document.getElementById('radio191'),
      radio192 = document.getElementById('radio192'),
      radio193 = document.getElementById('radio193'),
      radio201 = document.getElementById('radio201'),
      radio202 = document.getElementById('radio202'),
      radio203 = document.getElementById('radio203'),
      testform = document.getElementById('testform'),
      t1 = async_test("click on mutable radio fires click event, then input event, then change event"),
      t3 = async_test("click on non-mutable radio doesn't fire the input event"),
      t4 = async_test("click on non-mutable radio doesn't fire the change event"),
      t5 = async_test("canceled activation steps on unchecked radio"),
      input_fired = false,
      change_fired = false;

  test(function(){
    assert_false(radio1.checked);
    assert_false(radio2.checked);
    radio1.checked = true;
    assert_true(radio1.checked);
    assert_false(radio2.checked);
    radio2.checked = true;
    assert_false(radio1.checked);
    assert_true(radio2.checked);
  }, "only one control of a radio button group can have its checkedness set to true");

  test(function(){
    assert_false(radio3.checked);
    assert_false(radio4.checked);
    radio3.checked = true;
    assert_true(radio3.checked);
    assert_false(radio4.checked);
    radio4.checked = true;
    assert_false(radio3.checked);
    assert_true(radio4.checked);
  }, "radio inputs with non-ASCII name attributes belong to the same radio button group");

  test(function(){
    assert_true(radio8.checked);
    assert_false(radio9.checked);
    assert_false(radio10.checked);
    assert_true(radio11.checked);
    radio9.name="group4";
    radio9.checked = true;
    assert_true(radio8.checked);
    assert_true(radio9.checked);
    assert_false(radio10.checked);
    assert_false(radio11.checked);
  }, "changing the name of a radio input element and setting its checkedness to true makes all the other elements' checkedness in the same radio button group be set to false");

  test(function(){
    radio12.remove();
    assert_true(radio12.checked);
    assert_false(radio13.checked);
    assert_false(radio14.checked);
    radio13.checked = true;
    assert_true(radio13.checked);
    assert_false(radio14.checked);
    radio13.removeAttribute("form");
    radio14.removeAttribute("form");
    assert_true(radio13.checked);
    assert_false(radio14.checked);
    radio14.checked = true;
    assert_false(radio13.checked);
    assert_true(radio14.checked);
    radio13.setAttribute("form", "testform");
    radio14.setAttribute("form", "testform");
    radio13.checked = true;
    assert_true(radio13.checked);
    assert_false(radio14.checked);
    testform.remove();
    assert_true(radio13.checked);
    assert_false(radio14.checked);
  }, "moving radio input element out of or into a form should still work as expected");

  test(function(){
    assert_true(radio151.checked);
    assert_false(radio152.checked);
    assert_true(radio153.checked);
    assert_false(radio154.checked);
    assert_true(radio155.checked);
    assert_false(radio156.checked);
    radio152.checked = true;
    assert_false(radio151.checked);
    assert_true(radio152.checked);
    assert_true(radio153.checked);
    assert_false(radio154.checked);
    assert_true(radio155.checked);
    assert_false(radio156.checked);
    radio154.checked = true;
    assert_false(radio151.checked);
    assert_true(radio152.checked);
    assert_false(radio153.checked);
    assert_true(radio154.checked);
    assert_true(radio155.checked);
    assert_false(radio156.checked);
    radio156.checked = true;
    assert_false(radio151.checked);
    assert_true(radio152.checked);
    assert_false(radio153.checked);
    assert_true(radio154.checked);
    assert_false(radio155.checked);
    assert_true(radio156.checked);
  }, "different form owners and no form owner can't effect each other checkedness states");

  test(function(){
    assert_true(radio161.checked);
    assert_true(radio162.checked);
    assert_false(radio163.checked);
    radio163.checked = true;
    assert_false(radio161.checked);
    assert_true(radio162.checked);
    assert_true(radio163.checked);
    radio162.checked = false;
    assert_false(radio161.checked);
    assert_false(radio162.checked);
    assert_true(radio163.checked);
  }, "if input elements' type attribute are not radio, they can't effect each other checkedness states");

  test(function(){
    assert_true(radio171.checked);
    assert_false(radio172.checked);
    radio172.checked = true;
    assert_true(radio171.checked);
    assert_true(radio172.checked);
  }, "if the radio elements' name attribute are empty, they can't effect each other checkedness states");

  test(function(){
    assert_false(radio181.checked);
    assert_false(radio182.checked);
    var elements = document.querySelectorAll("#required :invalid");
    assert_array_equals(elements, document.getElementsByName("group9"));
    radio181.checked = true;
    assert_true(radio181.checked);
    assert_false(radio182.checked);
    elements = document.querySelectorAll("#required :invalid");
    assert_array_equals(elements, []);
    radio182.checked = true;
    assert_false(radio181.checked);
    assert_true(radio182.checked);
    elements = document.querySelectorAll("#required :valid");
    assert_array_equals(elements, document.getElementsByName("group9"));
  }, "test the constraint validation is correct when an element in the radio button group is required");

  test(function(){
    assert_true(radio191.checked);
    assert_true(radio192.checked);
    assert_false(radio193.checked);
    radio192.setAttribute("type", "radio");
    assert_false(radio191.checked);
    assert_true(radio192.checked);
    assert_false(radio193.checked);
  }, "changing an input element's type attribute to radio and setting its checkedness to true makes all the other elements' checkedness in the same radio button group be set to false");

  test(function(){
    assert_true(radio201.checked);
    assert_true(radio202.checked);
    assert_false(radio203.checked);
    radio201.setAttribute("form", "testform202");
    assert_true(radio201.checked);
    assert_false(radio202.checked);
    assert_false(radio203.checked);
  }, "changing the form owner of a radio input element and setting its checkedness to true makes all the other elements' checkedness in the same radio button group be set to false");

  radio5.onclick = t1.step_func(function(e) {
    click_fired = true;
    assert_false(input_fired, "click event should fire before input event");
    assert_false(change_fired, "click event should fire before change event");
    assert_false(e.isTrusted, "click()-initiated click event shouldn't be trusted");
  });

  radio5.oninput = t1.step_func(function(e) {
    input_fired = true;
    assert_true(click_fired, "input event should fire after click event");
    assert_false(change_fired, "input event should fire before change event");
    assert_true(e.bubbles, "input event should bubble")
    assert_true(e.isTrusted, "input event should be trusted");
    assert_false(e.cancelable, "input event should not be cancelable");
  });

  radio5.onchange = t1.step_func(function(e) {
    change_fired = true;
    assert_true(click_fired, "change event should fire after click event");
    assert_true(input_fired, "change event should fire after input event");
    assert_true(e.bubbles, "change event should bubble")
    assert_true(e.isTrusted, "change event should be trusted");
    assert_false(e.cancelable, "change event should not be cancelable");
  });

  radio6.oninput= t3.step_func_done(function(e) {
    assert_unreached("event input fired");
  });

  radio6.onchange = t4.step_func_done(function(e) {
    assert_unreached("event change fired");
  });

  t1.step(function() {
    radio5.click();
    assert_true(input_fired);
    t1.done();
  });

  t3.step(function(){
    radio6.click();
    t3.done();
    t4.done();
  });

  radio72.onclick = t5.step_func_done(function(e){
    assert_false(radio71.checked, "click on radio should uncheck other radio in same group");
    assert_true(radio72.checked, "click on radio should check that radio");
    e.preventDefault();
    // The cancelation of the click doesn't have an effect until after all the click event handlers have been run.
    assert_false(radio71.checked, "radio remains unchecked immediately after click event on other radio in same group is canceled");
    assert_true(radio72.checked, "clicked radio remains checked immediately after click event is canceled");
  });

  t5.step(function(){
    assert_true(radio71.checked, "initially checked radio should be checked");
    assert_false(radio72.checked, "other radios in same group as initially-checked radio should be unchecked");
    radio72.click();
    // Now that the click event has been fully dispatched, its cancelation has taken effect.
    assert_true(radio71.checked, "canceled click event on radio should leave the previously-checked radio checked");
    assert_false(radio72.checked, "canceled click event on previously-unchecked radio should leave that radio unchecked");
  });
</script>
